% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{compute_counts_annotation_data}
\alias{compute_counts_annotation_data}
\title{Compute number of values per group for count annotation}
\usage{
compute_counts_annotation_data(data, x, facet_rows, facet_columns)
}
\arguments{
\item{facet_columns}{}
}
\value{

}
\description{
Compute number of values per group for count annotation
}
