% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{compute_facet_dim}
\alias{compute_facet_dim}
\title{Get dimensions of ggplot plots with facets}
\usage{
compute_facet_dim(p)
}
\arguments{
\item{plot}{ggplot object with facets}
}
\value{

}
\description{
Get dimensions of ggplot plots with facets
}
\examples{
NULL
}
