% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{.add_group_numbers}
\alias{.add_group_numbers}
\title{Add group numbers to a pairwise_annotation with pairwise comparisons}
\usage{
.add_group_numbers(pairwise_annotation, data, x)
}
\arguments{
\item{pairwise_annotation}{data frame with pairwise annotation to be plotted, must have columns group1 and group2}
}
\value{

}
\description{
Adds columns group1_num, group2_num, and rank
}
\examples{
NULL
}
