% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{.add_tier_mapping}
\alias{.add_tier_mapping}
\title{Add tier mapping to pairwise annotation data based on tier mapping}
\usage{
.add_tier_mapping(pairwise_annotation, tier_mapping)
}
\arguments{
\item{pairwise_annotation}{data frame from .prepare_pairwise_annotation}

\item{tier_mapping}{result from .map_tiers}
}
\value{

}
\description{
Add tier mapping to pairwise annotation data based on tier mapping
}
\examples{
NULL
}
