% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{.assign_tiers}
\alias{.assign_tiers}
\title{Assign tiers based on optimal plotting location to prevent one tier per comparison}
\usage{
.assign_tiers(pairwise_annotation)
}
\arguments{
\item{pairwise_annotation}{data frame with pairwise annotation to be plotted, must have columns group1 and group2}
}
\value{

}
\description{
Assign tiers based on optimal plotting location to prevent one tier per comparison
}
\examples{
NULL
}
