% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{.assign_unassigned_tiers}
\alias{.assign_unassigned_tiers}
\title{Assigned tiers that have not been assigned yet}
\usage{
.assign_unassigned_tiers(pairwise_annotation)
}
\arguments{
\item{pairwise_annotation}{data frame with pairwise annotation to be plotted, must have columns group1_num and group2_num}
}
\value{

}
\description{
Assigned tiers that have not been assigned yet
}
