% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{.create_annotation}
\alias{.create_annotation}
\title{Create annotation for rows/columns}
\usage{
.create_annotation(data, which, show_legend, text_size, palette)
}
\arguments{
\item{data}{data frame with annotations}

\item{which}{either "row" or "column" for row annotations and column annotations respectively}

\item{show_legend}{columns of data to show legend for}

\item{text_size}{text size}

\item{palette}{color palette}
}
\value{

}
\description{
Create annotation for rows/columns
}
\examples{
NULL
}
