% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{.create_heatmap_color}
\alias{.create_heatmap_color}
\title{Create annotation for rows/columns}
\usage{
.create_heatmap_color(
  heatmap_color,
  matrix,
  lower_quantile,
  upper_quantile,
  color_by_quantile
)
}
\arguments{
\item{data}{data frame with annotations}

\item{which}{either "row" or "column" for row annotations and column annotations respectively}

\item{show_legend}{columns of data to show legend for}

\item{text_size}{text size}

\item{palette}{color palette}
}
\value{

}
\description{
Create annotation for rows/columns
}
\examples{
NULL
}
