% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.fix_outliers}
\alias{.fix_outliers}
\title{Remove or mask outliers based on quantiles in each group}
\usage{
.fix_outliers(data, y, lower_quantile, upper_quantile, drop_outliers, groups)
}
\arguments{
\item{data}{data frame containing dataset to use for plotting}

\item{lower_quantile}{lower quantile beyond which to limit axis}

\item{upper_quantile}{upper quantile beyond which to limit axis}

\item{drop_outliers}{whether to drop outliers or not - if FALSE, then outliers are masked to the lower or upper quantile values}

\item{groups}{columns which to group on in computing outliers}
}
\value{

}
\description{
Remove or mask outliers based on quantiles in each group
}
\keyword{internal}
