% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{.map_tiers}
\alias{.map_tiers}
\title{Map tiers to y positions based on the data}
\usage{
.map_tiers(data, pairwise_annotation, y, tier_width, scale)
}
\arguments{
\item{data}{data frame containing dataset on which the annotations will be plotted}

\item{pairwise_annotation}{result from .prepare_pairwise_annotations}

\item{y}{column for y-axis}

\item{tier_width}{relative distance between tiers for pairwise annotations}

\item{scale}{either "default" for linearly-spaced scale between y tier positions or "log" for log-spaced}
}
\value{

}
\description{
Map tiers to y positions based on the data
}
