% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{.plot_facet_grid}
\alias{.plot_facet_grid}
\title{Helper function to add facet_grid to plot}
\usage{
.plot_facet_grid(plot, facet_rows, facet_columns, ...)
}
\arguments{
\item{plot}{ggplot object to facet}

\item{facet_rows}{columns in data to make facet rows}

\item{facet_columns}{columns in data to make facet columns}
}
\value{

}
\description{
Helper function to add facet_grid to plot
}
\examples{
NULL
}
