% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{.plot_facet_wrap}
\alias{.plot_facet_wrap}
\title{Helper function to add facet_wrap to plot}
\usage{
.plot_facet_wrap(plot, facet_rows, facet_columns, ...)
}
\arguments{
\item{plot}{ggplot object to facet}

\item{facet_rows}{columns in data to facet}

\item{facet_columns}{same as facet_rows for facet_wrap - just combines these two with rows first then columns}
}
\value{

}
\description{
Helper function to add facet_wrap to plot
}
