% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.plot_jitter}
\alias{.plot_jitter}
\title{Plot jitter plot}
\usage{
.plot_jitter(data, x, y, color, fill, group, alpha, point_size)
}
\arguments{
\item{data}{Data frame for plotting}

\item{x}{Column for x-axis}

\item{y}{Column for y-axis}

\item{color}{Column to color points by}

\item{group}{Column to group points by - not relevant for this function}

\item{alpha}{Alpha for each point}
}
\value{

}
\description{
Plot jitter plot
}
\examples{
NULL
}
