% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.plot_line}
\alias{.plot_line}
\title{Plot line graph}
\usage{
.plot_line(data, x, y, color, fill, group, alpha, point_size)
}
\arguments{
\item{data}{data frame for plotting}

\item{x}{column for x-axis}

\item{y}{column for y-axis}

\item{color}{column to color points by}

\item{fill}{column to fill points by}

\item{group}{column to group points by - line connects by this variable}

\item{alpha}{alpha for each point}
}
\value{

}
\description{
Plot line graph
}
\examples{
NULL
}
