% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{.plot_overview_continuous_distribution}
\alias{.plot_overview_continuous_distribution}
\title{Function to plot distribution of continuous variable}
\usage{
.plot_overview_continuous_distribution(data, column, facet_rows, text_size = 3)
}
\description{
Function to plot distribution of continuous variable
}
