% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{.plot_overview_discrete_distribution}
\alias{.plot_overview_discrete_distribution}
\title{Function to plot distribution of discrete variable}
\usage{
.plot_overview_discrete_distribution(data, column, facet_rows, text_size = 3)
}
\description{
Function to plot distribution of discrete variable
}
