% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.plot_ridge}
\alias{.plot_ridge}
\title{Plot ridge plot}
\usage{
.plot_ridge(data, x = NULL, y = NULL, color = NULL, group = NULL, alpha = 0.5)
}
\arguments{
\item{data}{Data frame for plotting}

\item{x}{Column for x-axis}

\item{y}{Column for y-axis}

\item{color}{Column to color points by}

\item{group}{Column to group points by - not relevant for this function}

\item{alpha}{Alpha for each point}
}
\value{

}
\description{
Plot ridge plot
}
\examples{
NULL
}
