% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{.seq_log}
\alias{.seq_log}
\title{Sequence generation with equal distance in log space}
\usage{
.seq_log(from, by, length.out)
}
\arguments{
\item{from}{value to start from}

\item{by}{distance between points}

\item{length.out}{desired vector length}
}
\value{

}
\description{
Sequence generation with equal distance in log space
}
