% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{generate_palette_complex_heatmap}
\alias{generate_palette_complex_heatmap}
\title{Generate palette for ComplexHeatmap}
\usage{
generate_palette_complex_heatmap(annotations, palette)
}
\arguments{
\item{annotations}{data frame of annotations}

\item{palette}{list of character vectors, each element in list with the same name as columns in data frame, and each character vector containing the value as the name and color as the value.}
}
\value{

}
\description{
Takes in data frame of annotations and desired colors for each value in each column, and fills in missing values and returns palette in format required by ComplexHeatmap.
}
\examples{
NULL
}
