% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{generate_palette_ggplot}
\alias{generate_palette_ggplot}
\title{Generate palette for ggplot2}
\usage{
generate_palette_ggplot(values, palette)
}
\arguments{
\item{palette}{character vector containing the value as the name and color as the value.}

\item{annotations}{vector of values}
}
\value{

}
\description{
Takes in vector of values and desired colors for each value, and fills in missing values with default colors.
}
\examples{
NULL
}
