% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_palette}
\alias{get_palette}
\title{Generate ggexp palette with assigned color values based on alphabetized input values and overriden with input palette}
\usage{
get_palette(values, palette = c())
}
\arguments{
\item{values}{character vector or factor which to assign colors to}
}
\value{
ggplot color/fill scale object
}
\description{
Generate ggexp palette with assigned color values based on alphabetized input values and overriden with input palette
}
