% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{map_tiers}
\alias{map_tiers}
\title{Map tiers to y positions based on the data}
\usage{
map_tiers(data, pairwise_annotation, y, tier_width, scale)
}
\arguments{
\item{scale}{}
}
\description{
Map tiers to y positions based on the data
}
