% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_box}
\alias{plot_box}
\title{Plot box and whiskers plot}
\usage{
plot_box(data, x = NULL, y = NULL, color = NULL, group = NULL,
  alpha = 0.5)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot box and whiskers plot
}
