% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_counts_annotation}
\alias{plot_counts_annotation}
\title{Annotate number of values per group on plot}
\usage{
plot_counts_annotation(plot, x, counts_annotation, annotate_counts, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Annotate number of values per group on plot
}
