% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plot_data_overview}
\alias{plot_data_overview}
\title{Plot overview plot, showing distributions of all columns in a data frame,}
\usage{
plot_data_overview(data, columns, facet_rows, text_size)
}
\arguments{
\item{text_size}{}
}
\value{

}
\description{
Plot overview plot, showing distributions of all columns in a data frame,
}
