% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_density}
\alias{plot_density}
\title{Plot density plot}
\usage{
plot_density(data, x = NULL, y = NULL, color = NULL, group = NULL, alpha = 0.5)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot density plot
}
