% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_facets}
\alias{plot_facets}
\title{Utility function to facet a plot}
\usage{
plot_facets(plot, facet_rows, facet_columns, facet_type, facet_scales,
  facet_switch, nrow)
}
\arguments{
\item{plot}{ggplot object}

\item{facet_rows}{Column names from original data frame to be used in facet rows if facet_type == "grid" else columns to be faceted}

\item{facet_columns}{Column names from original data frame to be used in facet columns if facet_type == "grid" else columns to be faceted}

\item{facet_type}{Either "wrap" or "grid"}

\item{facet_scales}{Corresponds to scales argument in facet_wrap/facet_grid - either "fixed", free", "free_x", "free_y"}

\item{facet_switch}{Corresponds to switch argument in facet_wrap/facet_grid - either "x", "y", or "both"}

\item{nrow}{If facet_wrap, max number of rows in faceted figure}
}
\value{
ggplot object
}
\description{
Utility function to facet a plot
}
