% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap using ComplexHeatmap using simplified interface}
\usage{
plot_heatmap(
  matrix,
  row_annotations = NULL,
  column_annotations = NULL,
  split_rows = NULL,
  split_columns = NULL,
  palette = list(),
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  show_row_dend = TRUE,
  show_column_dend = TRUE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  show_legend_row = colnames(row_annotations),
  show_legend_column = colnames(column_annotations),
  legend_position = "side",
  lower_quantile = 0.01,
  upper_quantile = 0.99,
  text_size = 10,
  row_names_size = 4,
  column_names_size = 4,
  value_name = "value"
)
}
\arguments{
\item{matrix}{matrix to plot}

\item{row_annotations}{data frame with annotations for the rows of the matrix}

\item{column_annotations}{data frame with annotations for the columns of the matrix}

\item{split_rows}{columns of row_annotations to split heatmap rows by}

\item{split_columns}{columns of column_annottations to  split heatmap columns by}

\item{palette}{named list of named character vectors with the values being the colors if discrete, or in the form c(colors = XXX, breaks = XXX) if  continuous}

\item{cluster_rows}{boolean to cluster rows or not}

\item{cluster_columns}{boolean to cluster columns or not}

\item{show_row_dend}{boolean to show row dendrogram or not}

\item{show_column_dend}{boolean to show column dendrogram or not}

\item{show_row_names}{boolean to show row names or not}

\item{show_column_names}{boolean to show column names or  not}

\item{show_legend_row}{columns of row_annotations to show legend for}

\item{show_legend_column}{columns of column_annotations to show legend for}

\item{legend_position}{either "right" or "bottom"}

\item{lower_quantile}{lower quantile of data for colorbar}

\item{upper_quantile}{upper quantile of data for colorbar}

\item{text_size}{text size}

\item{row_names_size}{text size for row names}

\item{column_names_size}{text size for column names}
}
\value{

}
\description{
Plot heatmap using ComplexHeatmap using simplified interface
}
\examples{
NULL
}
