% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_line}
\alias{plot_line}
\title{Plot line graph}
\usage{
plot_line(data, x = NULL, y = NULL, color = NULL, group = NULL, alpha = 0.5)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot line graph
}
