% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{plot_pairwise_annotation}
\alias{plot_pairwise_annotation}
\title{Plot pairwise annotation}
\usage{
plot_pairwise_annotation(plot, pairwise_annotation, label = "p_signif",
  values_to_exclude = c(), tier_width = 0.15, scale = "default")
}
\arguments{
\item{plot}{ggplot object with discrete x-axis}

\item{pairwise_annotation}{data frame from prepare_pairwise_annotation}

\item{label}{string scalar indicating column for text annotation}
}
\value{
ggplot object
}
\description{
Plot pairwise annotation
}
