% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_scatterplot.R
\name{plot_pairwise_scatterplot}
\alias{plot_pairwise_scatterplot}
\title{Plot pairwise scatterplots}
\usage{
plot_pairwise_scatterplot(
  data,
  x = colnames(data \%>\% select_if(is.numeric)),
  y = x,
  combination_groups = c(),
  axis_annotations = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  alpha = 0.3,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  facet_scales = "free",
  facet_switch = "both",
  nrow = 2
)
}
\arguments{
\item{data}{Data frame to plot}

\item{x}{Columns to use for x-axis}

\item{y}{Columns to use for y-axis}

\item{color}{Column for color}

\item{shape}{Column for shape}

\item{alpha}{Numeric value between 0 and 1 for alpha}

\item{facet_rows}{Columns to facet on}

\item{facet_columns}{Columns to facet on}

\item{facet_type}{Either "wrap" or "grid", same as ggplot}

\item{facet_scales}{Either NULL, "fixed", "free", "free_x", "free_y", same as ggplot}

\item{facet_switch}{Either NULL, "x", "y", "both", same as ggplot}

\item{nrow}{Number of rows if facet_type is "wrap"}
}
\value{
ggplot object
}
\description{
Plot pairwise scatterplots
}
\examples{
NULL
}
