% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_quasirandom}
\alias{plot_quasirandom}
\title{Plot quasirandom plot}
\usage{
plot_quasirandom(
  data,
  x = NULL,
  y = NULL,
  color = NULL,
  group = NULL,
  alpha = 0.5
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot quasirandom plot
}
