% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_ridge}
\alias{plot_ridge}
\title{Plot ridge plot}
\usage{
plot_ridge(data, x = NULL, y = NULL, color = NULL, group = NULL,
  alpha = 0.5)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot ridge plot
}
