% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_scale}
\alias{plot_scale}
\title{Transform scale based on scale type and plot type}
\usage{
plot_scale(plot, scale, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Transform scale based on scale type and plot type
}
