% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_violin}
\alias{plot_violin}
\title{Plot violin plot}
\usage{
plot_violin(data, x = NULL, y = NULL, color = NULL, group = NULL,
  alpha = 0.5)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot violin plot
}
