% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{prepare_pairwise_annotation}
\alias{prepare_pairwise_annotation}
\title{Prepare pairwise annotation data frame for plotting by assigning tiers and y positions}
\usage{
prepare_pairwise_annotation(
  data,
  pairwise_annotation,
  x,
  y,
  label,
  values_to_exclude = c(),
  groups = c(),
  tier_width = 0.15,
  scale = "default"
)
}
\arguments{
\item{data}{data frame containing dataset on which the annotations will be plotted}

\item{pairwise_annotation}{data frame containing pairwise annotations (for example, pairwise statistical tests)}

\item{y}{column for y-axis}

\item{label}{column for text annotation}

\item{values_to_exclude}{values to not annotate (for example, "ns" for statistical tests)}

\item{groups}{columns to group by before assigning optimal y positions}

\item{tier_width}{relative distance between tiers for pairwise annotations}

\item{scale}{either "default" for linearly-spaced scale between y tier positions or "log" for log-spaced}
}
\value{

}
\description{
Prepare pairwise annotation data frame for plotting by assigning tiers and y positions
}
