% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{prepare_pairwise_annotation_single_group}
\alias{prepare_pairwise_annotation_single_group}
\title{Prepare pairwise annotation data frame for plotting by assigning tiers and y positions for a single group}
\usage{
prepare_pairwise_annotation_single_group(
  data,
  pairwise_annotation,
  x,
  y,
  label,
  values_to_exclude,
  tier_width,
  scale
)
}
\arguments{
\item{data}{data frame containing dataset on which the annotations will be plotted}

\item{pairwise_annotation}{data frame containing pairwise annotations (for example, pairwise statistical tests)}

\item{y}{column for y-axis}

\item{label}{column for text annotation}

\item{values_to_exclude}{values to not annotate (for example, "ns" for statistical tests)}

\item{tier_width}{relative distance between tiers for pairwise annotations}

\item{scale}{either "default" for linearly-spaced scale between y tier positions or "log" for log-spaced}

\item{groups}{columns to group by before assigning optimal y positions}
}
\value{

}
\description{
Prepare pairwise annotation data frame for plotting by assigning tiers and y positions for a single group
}
