% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_palette}
\alias{get_palette}
\title{Generate ggexp palette with assigned color values based on alphabetized input values and overriden with input palette}
\usage{
get_palette(values = NULL, palette = c(), return_ggplot_scale = TRUE)
}
\arguments{
\item{return_ggplot_scale}{}
}
\value{

}
\description{
Generate ggexp palette with assigned color values based on alphabetized input values and overriden with input palette
}
