% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Plot (stacked) barplot with annotations}
\usage{
plot_barplot(
  data,
  x,
  y,
  label = NULL,
  color = NULL,
  fill = NULL,
  stacked = TRUE,
  constant_height = TRUE,
  text_size = 3,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  facet_switch = NULL,
  facet_scales = "free",
  nrow = 1
)
}
\arguments{
\item{data}{data frame containing dataset to use for plotting}

\item{x}{column for x-axis}

\item{y}{column for y-axis}

\item{label}{column for text annotation}

\item{color}{column for color}

\item{stacked}{boolean whether result should be a stacked barplot or not}

\item{constant_height}{boolean whether stacked bars should have constant height (proportions plotted)}

\item{facet_rows}{columns for faceting by row}

\item{facet_columns}{columns for faceting by column}

\item{facet_type}{either "wrap" or "grid", corresponding to facet_wrap and facet_grid respectively}

\item{facet_switch}{either NULL, "both", "x", or "y", same as switch argument in facet calls}

\item{facet_scales}{either "fixed", "free_x", "free_y", or "free", same as scales argument in facet calls}

\item{nrow}{number of rows in plot, only applies if facet_type == "wrap"}
}
\value{
ggplot object
}
\description{
This is a convenience function wrapping around geom_bar, but it makes it easy to make both stacked and unstacked plots and adjust text annotations accordingly.
}
\examples{
library(dplyr)

data = mtcars \%>\%
  mutate(cyl = factor(cyl)) \%>\%
  group_by(cyl, am) \%>\%
  tally()

plot_barplot(
  data = data,
  x = "am",
  y = "n",
  color = "cyl",
  label = "cyl"
)
}
