% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions with pairwise annotations and flexibility}
\usage{
plot_distributions(data, pairwise_annotation = NULL,
  type = "quasirandom", x, y, group = NULL, color = NULL,
  alpha = 0.5, scale = "default", lower_quantile = 0,
  upper_quantile = 1, tier_width = 0.16, annotate_counts = TRUE,
  pairwise_annotation_label = "p_signif",
  pairwise_annotation_exclude = c(), facet_rows = c(),
  facet_columns = c(), facet_type = "wrap", facet_scales = "free",
  facet_switch = NULL, nrow = 1)
}
\arguments{
\item{data}{data frame containing dataset to use for plotting}

\item{pairwise_annotation}{data frame containing pairwise annotations}

\item{type}{string scaler indicating type of plot - can be "line", "sina", "quasirandom", "density", "violin", "box", or "ridge"}

\item{x}{string scalar indicating column for x-axis}

\item{y}{string scalar indicating column for y-axis}

\item{group}{string scalar indicating column for group aesthethic, used if type == "line"}

\item{color}{string scalar indicating column for color}

\item{alpha}{numeric scalar for alpha of points}

\item{scale}{string scalar that is either "default" for linearly-spaced scale or "log" for log-spaced}

\item{lower_quantile}{numeric scalar indicating lower quantile of values, beyond which data points are filtered from plot}

\item{upper_quantile}{numeric scalar indicating uower quantile of values, beyond which data points are filtered from plot}

\item{tier_width}{numeric scalar indicating relative distance between tiers for pairwise annotations}

\item{annotate_counts}{boolean scalar indicating whether to annotate counts per group or not}

\item{pairwise_annotation_label}{string scalar indicating column of pairwise_annotation data to use for annotation text}

\item{pairwise_annotation_exclude}{string vector indicating values to not annotate on pairwise annotations}

\item{facet_rows}{string vector indicating columns for faceting by row}

\item{facet_columns}{string vector indicating columns for faceting by column}

\item{facet_type}{string scalar that is either "wrap" or "grid", corresponding to facet_wrap and facet_grid respectively}

\item{facet_scales}{string scalar that is either "fixed", "free_x", "free_y", or "free", same as scales argument in facet calls}

\item{facet_switch}{string scalar that is either NULL, "both", "x", or "y", same as switch argument in facet calls}

\item{nrow}{numeric scalar indicating the number of rows in plot, only applies if facet_type == "wrap"}
}
\value{
ggplot object
}
\description{
Plot distributions with pairwise annotations and flexibility
}
\examples{

}
