% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions with pairwise annotations and flexibility}
\usage{
plot_distributions(
  data,
  pairwise_annotation = NULL,
  type = "quasirandom",
  x,
  y,
  group = NULL,
  color = NULL,
  alpha = 0.5,
  scale = "default",
  tier_width = 0.16,
  annotate_counts = TRUE,
  pairwise_annotation_label = "p_signif",
  pairwise_annotation_exclude = c(),
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "wrap",
  facet_scales = "free",
  facet_switch = NULL,
  nrow = 1
)
}
\arguments{
\item{data}{data frame containing dataset to use for plotting}

\item{pairwise_annotation}{data frame containing pairwise annotations}

\item{type}{type of plot - can be "line", "sina", "quasirandom", "density", "violin", "box", or "ridge"}

\item{x}{column for x-axis}

\item{y}{column for y-axis}

\item{group}{column for group aesthethic, used if type == "line"}

\item{color}{column for color}

\item{alpha}{alpha of points}

\item{scale}{either "default" for linearly-spaced scale or "log" for log-spaced}

\item{tier_width}{relative distance between tiers for pairwise annotations, between 0 and 1}

\item{annotate_counts}{boolean whether to annotate counts per group or not}

\item{pairwise_annotation_label}{column of pairwise_annotation data to use for annotation text}

\item{pairwise_annotation_exclude}{values to not annotate on pairwise annotations}

\item{facet_rows}{columns for faceting by row}

\item{facet_columns}{columns for faceting by column}

\item{facet_type}{either "wrap" or "grid", corresponding to facet_wrap and facet_grid respectively}

\item{facet_scales}{either "fixed", "free_x", "free_y", or "free", same as scales argument in facet calls}

\item{facet_switch}{either NULL, "both", "x", or "y", same as switch argument in facet calls}

\item{nrow}{number of rows in plot, only applies if facet_type == "wrap"}
}
\value{
ggplot object
}
\description{
Plot distributions with pairwise annotations and flexibility
}
\examples{

}
