% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{plot_facets}
\alias{plot_facets}
\title{Utility function to facet a plot}
\usage{
plot_facets(
  plot,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  ...
)
}
\arguments{
\item{plot}{ggplot object}

\item{facet_rows}{column names from original data frame to be used in facet rows if facet_type == "grid" else columns to be faceted}

\item{facet_columns}{column names from original data frame to be used in facet columns if facet_type == "grid" else columns to be faceted}

\item{facet_type}{either "wrap" or "grid"}

\item{...}{params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{
ggplot object
}
\description{
Utility function to facet a plot
}
