% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_scatterplot.R
\name{plot_pairwise_scatterplot}
\alias{plot_pairwise_scatterplot}
\title{Plot pairwise scatterplots}
\usage{
plot_pairwise_scatterplot(
  data,
  x = colnames(data \%>\% select_if(is.numeric)),
  y = x,
  combination_groups = c(),
  axis_annotations = NULL,
  color = NULL,
  shape = NULL,
  alpha = 1,
  point_size = 0.5,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  ...
)
}
\arguments{
\item{data}{Data frame to plot}

\item{x}{Columns to use for x-axis}

\item{y}{Columns to use for y-axis}

\item{color}{Column for color}

\item{shape}{Column for shape}

\item{alpha}{Numeric value between 0 and 1 for alpha}

\item{facet_rows}{Columns to facet on}

\item{facet_columns}{Columns to facet on}

\item{facet_type}{Either "wrap" or "grid", same as ggplot}

\item{...}{params passed into either facet_wrap or facet_grid, depending on facet_type parameter}
}
\value{
ggplot object
}
\description{
Plot pairwise scatterplots
}
