% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_scatterplot.R
\name{plot_pairwise_scatterplot}
\alias{plot_pairwise_scatterplot}
\title{Plot pairwise scatterplots}
\usage{
plot_pairwise_scatterplot(
  data,
  x = colnames(data \%>\% select_if(is.numeric)),
  y = x,
  combination_group = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  alpha = 0.3,
  facet_rows = c(),
  facet_columns = c(),
  facet_type = "grid",
  facet_scales = "free",
  facet_switch = "both",
  nrow = 2
)
}
\arguments{
\item{data}{data frame containing dataset to use for plotting}

\item{x}{string scalar indicating column for x-axis}

\item{y}{string scalar indicating column for y-axis}

\item{color}{string scalar indicating column for color}

\item{shape}{string scalar indicating column for shape}

\item{size}{string scalar indicating column for size}

\item{alpha}{numeric scalar for alpha of points}

\item{facet_rows}{string vector indicating columns for faceting by row}

\item{facet_columns}{string vector indicating columns for faceting by column}

\item{facet_type}{string scalar that is either "wrap" or "grid", corresponding to facet_wrap and facet_grid respectively}

\item{facet_scales}{string scalar that is either "fixed", "free_x", "free_y", or "free", same as scales argument in facet calls}

\item{facet_switch}{string scalar that is either NULL, "both", "x", or "y", same as switch argument in facet calls}

\item{nrow}{numeric scalar indicating the number of rows in plot, only applies if facet_type == "wrap"}
}
\value{
ggplot object
}
\description{
Plot pairwise scatterplots
}
