% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_sina}
\alias{plot_sina}
\title{Plot sina plot}
\usage{
plot_sina(data, x = NULL, y = NULL, color = NULL, group = NULL,
  alpha = 0.5)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Plot sina plot
}
