% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{prepare_pairwise_annotation}
\alias{prepare_pairwise_annotation}
\title{Prepare pairwise annotation data frame for plotting by assigning tiers and y positions}
\usage{
prepare_pairwise_annotation(data, pairwise_annotation, x, y, label,
  values_to_exclude = c(), groups = c(), tier_width = 0.15,
  scale = "default")
}
\arguments{
\item{data}{data frame containing dataset on which the annotations will be plotted}

\item{pairwise_annotation}{pairwise annotations}

\item{scale}{}
}
\description{
Prepare pairwise annotation data frame for plotting by assigning tiers and y positions
}
