% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_annotation.R
\name{prepare_pairwise_annotation_single_group}
\alias{prepare_pairwise_annotation_single_group}
\title{Prepare pairwise annotation data frame for plotting by assigning tiers and y positions for a single group}
\usage{
prepare_pairwise_annotation_single_group(data, pairwise_annotation, x, y,
  label, values_to_exclude, tier_width, scale)
}
\arguments{
\item{scale}{}
}
\description{
Prepare pairwise annotation data frame for plotting by assigning tiers and y positions for a single group
}
